backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.5
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pusher==3.3.3
requests==2.31.0
gunicorn==21.2.0
setuptools

